src/requirements.txt
folium
googlemaps
pandas
pyppeteer
python-dotenv
python_docx
streamlit
altair
PyPDF2
PyMySQL
cryptography
polyline
st-supabase-connection
